% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coefs.R
\name{generate_coefs}
\alias{generate_coefs}
\title{Title}
\usage{
generate_coefs(data, covars, modeltype)
}
\arguments{
\item{data}{A dataframe used to create model}

\item{covars}{List of covariates to be included in model}

\item{modeltype}{Type of model (logistic, survival)}
}
\value{
dataframe
}
\description{
Title
}
\examples{
# add example here
}
