% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_final_results.R
\name{push_final_results}
\alias{push_final_results}
\title{push_final_results}
\usage{
push_final_results(
  source_repo,
  source_subdir = NULL,
  dest_repo,
  dest_subdir = NULL,
  source_files,
  commit_message = paste("source repo to destination repo", Sys.Date())
)
}
\arguments{
\item{source_repo}{The file path for the git repository that is the source of the files}

\item{source_subdir}{The name of the subdirectory in which the source files are located. Default is NULL}

\item{dest_repo}{The file path for the destination git repository}

\item{dest_subdir}{The name of the subdirectory into which the source files should be moved. Default is NULL}

\item{source_files}{A vector of files to be moved from the source repository to the destination repository}

\item{commit_message}{Message to attach to the commits for both source and destination repository (optional)}
}
\value{

}
\description{
This function takes a source repository, a destination repository, and a list of files to be copied
It commits and pushes any changes to the source repository, and then copies the files from the
source repository to the destination repository. After files are copied, they are staged, committed and pushed to the
destination repository.
}
\examples{

}
