% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coefs.R
\name{generate_coefs}
\alias{generate_coefs}
\title{generate_coefs}
\usage{
generate_coefs(
  data,
  outcome,
  covars = NULL,
  id = "id",
  model_type = c("linear", "logistic", "survival", "quantile"),
  quantile_tau = c(0.25, 0.5, 0.75),
  model_weights = NULL,
  labels = NULL
)
}
\arguments{
\item{data}{Data that will be used to create model}

\item{outcome}{The outcome for the model}

\item{covars}{A vector of covariate names. If NULL, a null model (no covariates) will be created.}

\item{id}{Name of the variable containing unique identifier}

\item{model_type}{Type of model (linear, logistic, survival, quantile)}

\item{quantile_tau}{Vector of desired centiles for quantile model}

\item{model_weights}{Name of variable containing model weights}

\item{labels}{Labels corresponding to covariates in model}
}
\value{
dataframe
}
\description{
This function takes a dataframe, a list of covariates and a model type to
create a model and save out the coefficients for that model
}
\examples{
# add example here
}
