% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pred.R
\name{generate_pred}
\alias{generate_pred}
\title{generate_pred}
\usage{
generate_pred(
  data,
  coefficients,
  model_type = c("linear", "logistic", "survival", "quantile"),
  id = "id",
  covariate = "covariate",
  value = "value",
  outcometime = 5,
  starttime = 0
)
}
\arguments{
\item{data}{Data for which predictions will be generated}

\item{coefficients}{Dataframe containing model coefficients}

\item{model_type}{Type of model (linear, logistic, survival, quantile)}

\item{id}{Name of the variable containing unique identifier. Default is "id"}

\item{covariate}{Name of the variable in "coefficients" data containing covariate name. Default is "covariate"}

\item{value}{Name of the variable in the "coefficients" data containing coefficient value. Default is "value"}

\item{outcometime}{For survival models, the desired timepoint for prediction, in whatever unit was used for time to event in creating model. Default is 5 years.}

\item{starttime}{For survival models, the starting timepoint for prediction, in whatever unit was used for time to event in creating model. Default is 0 years (time of surgery).}
}
\value{
dataframe
}
\description{
This function will take in two dataframes (cohort data and coefficients) and calculate functions for all patients with available data
This function generates the following variables:
\code{pred_xb}: For linear and quantile regression, this is the predicted value. For logistic and survival analysis, this is the linear predictor.
\code{event_pr}: For logistic and survival analysis, this is the predicted probability of the event (at the specified timepoint, for survival models). For linear and quantile regression, this is not generated.
\code{nonevent_pr}: For logistic and survival analysis, this is the predicted probability of not having the event (survival probability at the specified timepoint, for survival models). For linear and quantile regression, this is not generated.
}
\examples{
# add example here
}
