% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomo.utils-package.R
\docType{package}
\name{nomo.utils-package}
\alias{nomo.utils}
\alias{nomo.utils-package}
\title{nomo.utils: Utility Functions for nomo Package}
\description{
This package contains utility functions related to the nomo package.
}
\author{
\strong{Maintainer}: Emily Vertosick \email{vertosie@mskcc.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
